feat: display stats in node info
